Section3-CloudBuild-CICD/cloudrun-app/application-files/requirements.txt
google-cloud-bigquery==2.34.4
Flask==2.1.0
gunicorn==20.1.0
pandas
pyarrow
pytest